pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maximus-code-agent"
version = "1.0.0"
description = "Maximus Code Agent – a local-first AI coding agent with safety rails"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "httpx>=0.25",
    "openai>=1.10",
    "pytest>=7.4",
]

[project.optional-dependencies]
pg = ["psycopg[binary]>=3.1", "pgvector>=0.2"]
telegram = ["python-telegram-bot>=20.0"]
all = ["maximus-code-agent[pg,telegram]"]

[project.scripts]
mca = "mca.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mca"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
